pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "training-templates"
version = "0.0.1"
description = "A collection of templated workflows for machine learning model training."
readme = "README.md"
dependencies = ["pyyaml"]
requires-python = ">=3.7"

# https://pypi.org/project/pip-tools/
# https://hynek.me/til/pip-tools-and-pyproject-toml/
[project.optional-dependencies]
local = ["pyspark>=3.0", "delta-spark>=2.0", "scikit-learn>=1.0", "pandas>=1.0", "mlflow>=2.0", "xgboost>=1.0", "hyperopt"]
test = ["pytest>=7.0", "coverage[toml]", "pytest-cov", "isort", "black", "mypy"]

[tool.mypy]
ignore_missing_imports = true